added pxe class
